- Changed modelica_integer to long, so it is 64-bit on 64-bit platforms
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@7541 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ c_runtime/modelica_string.c
@@ -86,9 +86,9 @@ modelica_string_const modelica_integer_to_modelica_string(modelica_integer i, mo
 	char buf[400];
 	formatStr[0]='%';
 	if (leftJustified) {
-		sprintf(&formatStr[1],"-%dd",minLen);
+		sprintf(&formatStr[1],"-%dd",(int)minLen);
 	} else {
-		sprintf(&formatStr[1],"%dd",minLen);
+		sprintf(&formatStr[1],"%dd",(int)minLen);
 	}
 	sprintf(buf,formatStr,i);
 	return init_modelica_string(buf);
@@ -102,9 +102,9 @@ modelica_string_const modelica_real_to_modelica_string(modelica_real r,modelica_
 	char buf[400];
 	formatStr[0]='%';
 	if (leftJustified) {
-		sprintf(&formatStr[1],"-%d.%dg",minLen,signDigits);
+		sprintf(&formatStr[1],"-%d.%dg",(int)minLen,(int)signDigits);
 	} else {
-		sprintf(&formatStr[1],"%d.%dg",minLen,signDigits);
+		sprintf(&formatStr[1],"%d.%dg",(int)minLen,(int)signDigits);
 	}
 	sprintf(buf,formatStr,r);
 	return init_modelica_string(buf);
